+++ apps/cli/src/live-common-setup-base.ts
@@ -90,7 +90,9 @@ setSupportedCurrencies([
   "telos_evm",
   "coreum",
   "injective",
+  "casper",
   "neon_evm",
+  "lukso",
 ]);
 
 for (const k in process.env) setEnvUnsafe(k as EnvName, process.env[k]);