+++ apps/ledger-live-desktop/jest.config.js
@@ -37,12 +37,7 @@ const defaultConfig = {
 };
 
 module.exports = {
-  collectCoverage: true,
-  collectCoverageFrom: [
-    "src/**/*.{ts,tsx}",
-    "!src/**/*.test.{ts,tsx}",
-    "!src/**/*.spec.{ts,tsx}",
-  ],
+  collectCoverageFrom: ["src/**/*.{ts,tsx}", "!src/**/*.test.{ts,tsx}", "!src/**/*.spec.{ts,tsx}"],
   coverageReporters: ["json", "lcov", "json-summary"],
   projects: [
     {