+++ apps/ledger-live-desktop/src/renderer/components/Onboarding/Screens/SelectUseCase/index.tsx
@@ -3,6 +3,7 @@ import { useHistory } from "react-router-dom";
 import { useTranslation, Trans } from "react-i18next";
 import { Flex, Text } from "@ledgerhq/react-ui";
 import { DeviceModelId } from "@ledgerhq/devices";
+import { useFeature, isRecoverDisplayed } from "@ledgerhq/live-common/featureFlags/index";
 import { useDispatch } from "react-redux";
 import styled from "styled-components";
 import { UseCaseOption } from "./UseCaseOption";
@@ -22,7 +23,6 @@ import setupNanoDark from "./assets/setupNanoDark.png";
 import restoreUsingRecoverDark from "./assets/restoreUsingRecoverDark.png";
 
 import Illustration from "~/renderer/components/Illustration";
-import { FeatureToggle } from "@ledgerhq/live-common/featureFlags/index";
 import { openModal } from "~/renderer/actions/modals";
 
 registerAssets([
@@ -88,6 +88,7 @@ export function SelectUseCase({ setUseCase, setOpenedPedagogyModal }: Props) {
   const { deviceModelId } = useContext(OnboardingContext);
   const history = useHistory();
   const dispatch = useDispatch();
+  const servicesConfig = useFeature("protectServicesDesktop");
 
   return (
     <ScrollArea withHint>
@@ -187,7 +188,7 @@ export function SelectUseCase({ setUseCase, setOpenedPedagogyModal }: Props) {
                 );
               }}
             />
-            <FeatureToggle featureId="protectServicesDesktop">
+            {isRecoverDisplayed(servicesConfig, deviceModelId) && (
               <UseCaseOption
                 dataTestId="v3-onboarding-restore-using-recover"
                 id="restore-device"
@@ -214,7 +215,7 @@ export function SelectUseCase({ setUseCase, setOpenedPedagogyModal }: Props) {
                   }
                 }}
               />
-            </FeatureToggle>
+            )}
           </RightColumn>
         </Row>
       </SelectUseCaseContainer>