+++ apps/ledger-live-desktop/src/renderer/families/casper/AccountSubHeader.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import AccountSubHeader from "../../components/AccountSubHeader/index";
+
+export default function CasperAccountSubHeader() {
+  return <AccountSubHeader family="Casper" team="Zondax"></AccountSubHeader>;
+}