+++ apps/ledger-live-desktop/src/renderer/screens/exchange/Swap2/Form/SwapWebView.tsx
@@ -1,4 +1,3 @@
-import { usePageState } from "@ledgerhq/live-common/exchange/swap/hooks/index";
 import React, { useEffect, useMemo, useRef, useState } from "react";
 import { useSelector } from "react-redux";
 import styled from "styled-components";
@@ -11,13 +10,15 @@ import useTheme from "~/renderer/hooks/useTheme";
 import { Web3AppWebview } from "~/renderer/components/Web3AppWebview";
 import { WebviewAPI, WebviewState } from "~/renderer/components/Web3AppWebview/types";
 import { initialWebviewState } from "~/renderer/components/Web3AppWebview/helpers";
-import { WalletAPICustomHandlers } from "@ledgerhq/live-common/wallet-api/types";
 import { handlers as loggerHandlers } from "@ledgerhq/live-common/wallet-api/CustomLogger/server";
-import { setStoreValue } from "~/renderer/store";
 import { TopBar } from "~/renderer/components/WebPlatformPlayer/TopBar";
 
-type SwapWebProps = {
-  inputs: Partial<{
+type CustomHandlersParams<Params> = {
+  params: Params;
+};
+
+export type SwapWebProps = {
+  swapState?: Partial<{
     provider: string;
     fromAccountId: string;
     toAccountId: string;
@@ -26,20 +27,23 @@ type SwapWebProps = {
     rate: string;
     feeStrategy: string;
     customFeeConfig: string;
+    cacheKey: string;
+    loading: boolean;
+    error: boolean;
   }>;
-  pageState: ReturnType<typeof usePageState>;
+  redirectToProviderApp(_: string): void;
 };
 
 export const SWAP_WEB_MANIFEST_ID = "swap-live-app-demo-0";
 
-const SwapWebAppWrapper = styled.div(
-  () => `
-  height: 0px;
-  width: 0px;
+const SwapWebAppWrapper = styled.div<{ isDevelopment: boolean }>(
+  ({ isDevelopment }) => `
+  ${!isDevelopment ? "height: 0px;" : ""}
+  width: 100%;
 `,
 );
 
-const SwapWebView = ({ pageState, inputs }: SwapWebProps) => {
+const SwapWebView = ({ swapState, redirectToProviderApp }: SwapWebProps) => {
   const {
     colors: {
       palette: { type: themeType },
@@ -55,25 +59,34 @@ const SwapWebView = ({ pageState, inputs }: SwapWebProps) => {
   const manifest = localManifest || remoteManifest;
 
   const hasManifest = !!manifest;
-  const hasInputs = !!inputs;
-  const isPageStateLoaded = pageState === "loaded";
+  const hasSwapState = !!swapState;
 
-  const customHandlers = useMemo<WalletAPICustomHandlers>(() => {
+  const customHandlers = useMemo(() => {
     return {
       ...loggerHandlers,
-      "storage.set": ({
-        params: { key, value },
-      }: {
-        params: { key: string; value: SwapLiveError };
-      }) => {
-        if (key === "error" && value.origin === "swap-web-app") {
-          onSwapWebviewError(value);
-        }
-        setStoreValue(key, {}, SWAP_WEB_MANIFEST_ID);
+      "custom.swapStateGet": () => {
+        return Promise.resolve(swapState);
+      },
+      // TODO: when we need bidirectional communication
+      // "custom.swapStateSet": (params: CustomHandlersParams<unknown>) => {
+      //   return Promise.resolve();
+      // },
+      "custom.throwExchangeErrorToLedgerLive": ({
+        params,
+      }: CustomHandlersParams<SwapLiveError>) => {
+        onSwapWebviewError(params);
+        return Promise.resolve();
+      },
+      "custom.throwGenericErrorToLedgerLive": () => {
+        onSwapWebviewError();
+        return Promise.resolve();
+      },
+      "custom.redirectToProviderApp": ({ params }: { params: string }) => {
+        redirectToProviderApp(params);
       },
     };
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  }, [swapState?.cacheKey]);
 
   useEffect(() => {
     if (webviewState.url.includes("/unknown-error")) {
@@ -83,7 +96,7 @@ const SwapWebView = ({ pageState, inputs }: SwapWebProps) => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [webviewState.url]);
 
-  if (!hasManifest || !hasInputs || !isPageStateLoaded) {
+  if (!hasManifest || !hasSwapState) {
     return null;
   }
   const onSwapWebviewError = (error?: SwapLiveError) => {
@@ -97,18 +110,18 @@ const SwapWebView = ({ pageState, inputs }: SwapWebProps) => {
       {isDevelopment && (
         <TopBar manifest={manifest} webviewAPIRef={webviewAPIRef} webviewState={webviewState} />
       )}
-      <SwapWebAppWrapper>
+      <SwapWebAppWrapper isDevelopment={isDevelopment}>
         <Web3AppWebview
           manifest={manifest}
           inputs={{
-            ...inputs,
+            cacheKey: swapState.cacheKey,
             theme: themeType,
             lang: locale,
             currencyTicker: fiatCurrency.ticker,
           }}
           onStateChange={setWebviewState}
           ref={webviewAPIRef}
-          customHandlers={customHandlers}
+          customHandlers={customHandlers as never}
         />
       </SwapWebAppWrapper>
     </>