+++ apps/ledger-live-desktop/tools/main.js
@@ -135,13 +135,6 @@ const build = async argv => {
     ),
   ]);
 
-  // Ensure that we keep our bundle size under thresholds
-  if (results[0].metafile.outputs[".webpack/main.bundle.js"].bytes > 5 * 1024 * 1024) {
-    throw new Error(
-      "main bundle must be kept under 5 MB. This indicates a possible regression of importing too much modules. Most of Ledger Live must be run on renderer side.",
-    );
-  }
-
   if (process.env.GENERATE_METAFILES) {
     // analyze bundle sizes. use it with https://esbuild.github.io/analyze/
     fs.writeFileSync("metafile.main.json", JSON.stringify(results[0].metafile), "utf-8");