+++ apps/ledger-live-mobile/src/components/RootNavigator/types/SendFundsNavigator.ts
@@ -34,6 +34,7 @@ import type { Transaction as RippleTransaction } from "@ledgerhq/live-common/fam
 import type { Transaction as ICPTransaction } from "@ledgerhq/live-common/families/internet_computer/types";
 import type { Transaction as StellarTransaction } from "@ledgerhq/live-common/families/stellar/types";
 import type { Transaction as StacksTransaction } from "@ledgerhq/live-common/families/stacks/types";
+import type { Transaction as CasperTransaction } from "@ledgerhq/live-common/families/casper/types";
 import BigNumber from "bignumber.js";
 import { Result } from "@ledgerhq/live-common/bridge/useBridgeTransaction";
 import { ScreenName } from "../../../const";
@@ -171,7 +172,6 @@ export type SendFundsNavigatorStackParamList = {
     transaction: EvmTransaction;
     setTransaction: Result<EvmTransaction>["setTransaction"];
     gasOptions?: GasOptions;
-    goBackOnSetTransaction?: boolean;
     setCustomStrategyTransactionPatch: React.Dispatch<
       React.SetStateAction<Partial<EvmTransaction> | undefined>
     >;
@@ -321,6 +321,20 @@ export type SendFundsNavigatorStackParamList = {
     account: Account;
     transaction: StacksTransaction;
     memoType?: string;
+    currentNavigation:
+      | ScreenName.SignTransactionSummary
+      | ScreenName.SendSummary
+      | ScreenName.SwapForm;
+    nextNavigation:
+      | ScreenName.SignTransactionSelectDevice
+      | ScreenName.SendSelectDevice
+      | ScreenName.SwapForm;
+  };
+  [ScreenName.CasperEditTransferId]: {
+    accountId: string;
+    account: Account;
+    parentId?: string;
+    transaction: CasperTransaction;
     currentNavigation:
       | ScreenName.SignTransactionSummary
       | ScreenName.SendSummary