+++ apps/ledger-live-mobile/src/components/TabBar/TransferDrawer.tsx
@@ -39,7 +39,7 @@ type ButtonItem = {
 export default function TransferDrawer({ onClose }: Omit<ModalProps, "isRequestingToBeOpened">) {
   const navigation = useNavigation();
   const {
-    quickActionsList: { SEND, RECEIVE, BUY, SELL, SWAP, STAKE, WALLET_CONNECT, RECOVER },
+    quickActionsList: { SEND, RECEIVE, BUY, SELL, SWAP, STAKE, RECOVER },
   } = useQuickActions();
   const { t } = useTranslation();
   const { pushToast, dismissToast } = useToasts();
@@ -112,7 +112,6 @@ export default function TransferDrawer({ onClose }: Omit<ModalProps, "isRequesti
       },
       title: t("transfer.buy.title"),
       description: t("transfer.buy.description"),
-      tag: t("common.popular"),
       Icon: BUY.icon,
       onPress: () => onNavigate(...BUY.route),
       onDisabledPress: () => {
@@ -182,6 +181,7 @@ export default function TransferDrawer({ onClose }: Omit<ModalProps, "isRequesti
       title: t("transfer.swap.title"),
       description: t("transfer.swap.description"),
       Icon: SWAP.icon,
+      tag: t("common.popular"),
       onPress: () => onNavigate(...SWAP.route),
       onDisabledPress: () => {
         if (isPtxServiceCtaExchangeDrawerDisabled) {
@@ -214,24 +214,6 @@ export default function TransferDrawer({ onClose }: Omit<ModalProps, "isRequesti
             onPress: () => onNavigateRecover(),
             disabled: RECOVER.disabled,
             testID: "transfer-recover-button",
-          },
-        ]
-      : []),
-
-    ...(WALLET_CONNECT
-      ? [
-          {
-            eventProperties: {
-              button: "transfer_walletConnect",
-              page,
-              drawer: "trade",
-            },
-            title: t("transfer.walletConnect.title"),
-            description: t("transfer.walletConnect.description"),
-            Icon: WALLET_CONNECT.icon,
-            onPress: () => onNavigate(...WALLET_CONNECT.route),
-            disabled: WALLET_CONNECT.disabled,
-            testID: "transfer-walletconnect-button",
           },
         ]
       : []),