+++ libs/ledger-live-common/src/families/casper/speculos-deviceActions.ts
@@ -1,0 +1,54 @@
+import type { DeviceAction } from "../../bot/types";
+import type { Transaction } from "./types";
+import { deviceActionFlow, SpeculosButton } from "../../bot/specs";
+import { casperPubKeyToAccountHash } from "./bridge/bridgeHelpers/addresses";
+
+export const acceptTransaction: DeviceAction<Transaction, any> = deviceActionFlow({
+  steps: [
+    {
+      title: "Please",
+      button: SpeculosButton.RIGHT,
+      expectedValue: () => "review",
+    },
+    {
+      title: "Txn hash",
+      button: SpeculosButton.RIGHT,
+    },
+    {
+      title: "Type",
+      button: SpeculosButton.RIGHT,
+      expectedValue: () => "Token transfer",
+    },
+    {
+      title: "Chain ID",
+      button: SpeculosButton.RIGHT,
+      expectedValue: () => "casper",
+    },
+    {
+      title: "Account",
+      button: SpeculosButton.RIGHT,
+      expectedValue: ({ account }) => account.freshAddress,
+    },
+    {
+      title: "Fee",
+      button: SpeculosButton.RIGHT,
+      expectedValue: ({ transaction }) =>
+        `${transaction.fees.toNumber().toLocaleString("en-US").replaceAll(",", " ")} motes`,
+    },
+    {
+      title: "Target",
+      button: SpeculosButton.RIGHT,
+      expectedValue: ({ transaction }) => casperPubKeyToAccountHash(transaction.recipient),
+    },
+    {
+      title: "Amount",
+      button: SpeculosButton.RIGHT,
+      expectedValue: ({ status }) =>
+        `${status.amount.toNumber().toLocaleString("en-US").replaceAll(",", " ")} motes`,
+    },
+    {
+      title: "APPROVE",
+      button: SpeculosButton.BOTH,
+    },
+  ],
+});