+++ tools/github-bot/src/features/autoClose/tools.ts
@@ -16,20 +16,17 @@ export const isValidUser = (user: string): boolean =>
 export const isValidBody = (body: string | null): boolean => {
   if (!body) return false;
 
-  const description =
-    "_Replace this text by a clear and concise description of what this pull request is about and why it is needed._";
-
   const requiredHeadings = [
     "### 📝 Description",
     "### ❓ Context",
     "### ✅ Checklist",
-    // "### 🚀 Expectations to reach",
+    "### 🧐 Checklist for the PR Reviewers",
   ];
 
   const results = body.split(/\r?\n/).reduce(
     (acc, line) => {
       // Dummy description line has not been replaced.
-      if (line === description) {
+      if (line.includes("_Replace this text")) {
         return {
           ...acc,
           dummyDescription: true,